+++ src/GraphQL.Tests/DI.xUnit/DependencyInjectionAdapterData.cs
@@ -1,6 +1,5 @@
 using System.Collections;
 using GraphQL.DI;
-using Microsoft.Extensions.DependencyInjection;
 
 namespace GraphQL.Tests.DI;
 
@@ -21,22 +20,4 @@ internal sealed class DependencyInjectionAdapterData : IEnumerable<object[]>
     }
 
     IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
-}
-
-internal interface IDependencyInjectionAdapter
-{
-    public IServiceProvider ServiceProvider { get; }
-}
-
-internal sealed class MicrosoftAdapter : IDependencyInjectionAdapter
-{
-    private readonly MicrosoftDI.GraphQLBuilder _builder;
-    private IServiceProvider _provider;
-
-    public MicrosoftAdapter(Action<IServiceRegister> configure)
-    {
-        _builder = new MicrosoftDI.GraphQLBuilder(new ServiceCollection(), b => configure(b.Services));
-    }
-
-    public IServiceProvider ServiceProvider => _provider ??= _builder.ServiceCollection.BuildServiceProvider();
 }
