+++ src/GraphQL.Tests/DI.xUnit/IDependencyInjectionAdapter.cs
@@ -1,0 +1,6 @@
+﻿namespace GraphQL.Tests.DI;
+
+internal interface IDependencyInjectionAdapter
+{
+    public IServiceProvider ServiceProvider { get; }
+}
