+++ src/GraphQL.Tests/StarWars/StarWarsTestBase.cs
@@ -9,13 +9,13 @@ namespace GraphQL.Tests.StarWars
         public override void RegisterServices(IServiceRegister Services)
         {
             Services.Singleton(new StarWarsData());
-            Services.Register<StarWarsQuery>();
-            Services.Register<StarWarsMutation>();
-            Services.Register<HumanType>();
-            Services.Register<HumanInputType>();
-            Services.Register<DroidType>();
-            Services.Register<CharacterInterface>();
-            Services.Register<EpisodeEnum>();
+            Services.Transient<StarWarsQuery>();
+            Services.Transient<StarWarsMutation>();
+            Services.Transient<HumanType>();
+            Services.Transient<HumanInputType>();
+            Services.Transient<DroidType>();
+            Services.Transient<CharacterInterface>();
+            Services.Transient<EpisodeEnum>();
 
             Services.Singleton<StarWarsSchema>();
         }
